Big re-factor of code to match what i want to do with standards for openGL, clen up of main and re-structure of directories.
naming same as documentation examples
re-structure and single shader parse
texture test added, call bound to 1 the issue was the vert shader was not linked to the same bind as the fragment shader so we never could find the parameter in the current shader program. Making this one file is a temporary solution.